+++ packages/fusion-runtime/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @graphql-mesh/fusion-runtime
 
+## 1.8.4
+### Patch Changes
+
+- Updated dependencies []:
+  - @graphql-tools/federation@4.3.4
+
 ## 1.8.3
 ### Patch Changes